pyproject.toml
[tool.poetry]
name = "orangescogs"
version = "0.1.0"
description = "Utility cogs for tgstation codebases"
authors = ["oranges <email@oranges.net.nz>", "crossedfall <https://github.com/Crossedfall>", "Will Roberts <https://github.com/wroberts"]
maintainers = ["oranges <email@oranges.net.nz>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">3.8.1,<3.10"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
Red-DiscordBot = "^3.4.16"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
